Main commit
first loaded in the song mp3 and added sound related variables for volume and frequencies. 

Then changes the moveAndDrawCircles function so that the current volume affects the opacity of the circles

Since browsers block autoplaying music, I added text prompting the user to click if music is not playing and a variable keeping track of whether there is music playing. I created an onMouseCLick function to play the song.

I introduced a smoothVolume functioning to make sure there was no flickering from volume changes.


I decided to have the circles and screamer become progressively more red and reduced green and blue to ensure it did not become too bright, enhancing Munches message. The low frequencies cause the circles to start changing colour, and also gain speed.
Furthermore, I modified the onMouseClick function to reset colour adjustments.

I then made it so that the high  frequencies (the wailing in the music) cause the screamers mouth to start screaming,